feat: 수익률 출력 구현

+++ src/main/java/lotto/util/ProfitCalculator.java
@@ -2,7 +2,7 @@ package lotto.util;
 
 public class ProfitCalculator {
 
-    public static double calculateProfitRate(int investmentAmount, int profit) {
+    public static double calculateProfitRate(int investmentAmount, long profit) {
         return Math.ceil(profit * 100 / investmentAmount) / 100;
     }
